% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/designPool.R
\name{designPool}
\alias{designPool}
\title{Generate design pool}
\usage{
designPool(type = NULL, traj = NULL, time = NULL, budget = NULL,
  unitcost = NULL, costratio = 1L, nfix = NULL, attrition = NULL, ...)
}
\arguments{
\item{type}{A character string indicating the design type.
Options include "com" (for complete data designs only),
"miss" (for planned missing data designs only), and "both".}

\item{traj}{A character string defining the shape of trajectory.
Options include "linear" and "quadratic".}

\item{time}{The number of maximum possible measurement occasions.
If \code{type} = "miss" or "both",
the maximum number of \code{time} is 6;
if \code{type} = "com",
the maximum number of \code{time} is 10.
If \code{traj} = "linear",
the minimum number of \code{time} is 3;
if \code{traj} = "quadratic",
the minimum number of \code{time} is 4.}

\item{budget}{The total amount of budget for the longitudinal study.
It is used together with \code{unitcost} to determine the
sample size (N) for each design.}

\item{unitcost}{The cost of a single measurement for each participant.
It is used together with \code{budget} to determine the
sample size (N) for each design.}

\item{costratio}{The cost ratio, indicating how many times more
it costs by recruiting a new participants than
a repeated measure (\code{unitcost}). The default is 1.
Given \code{budget}, \code{unitcost} and \code{costratio},
the sample size (N) can be computed as
\code{budget/(costratio * unitcost + (time - 1) * unitcost)}.}

\item{nfix}{The fixed sample size.
Used only when the sample size is fixed across different designs,
DO NOT use \code{nfix} together with
\code{budget}, \code{unitcost} and \code{costratio},
the sample size determined by \code{budget}, \code{unitcost}
and \code{costratio}, will overwrite \code{nfix}. The default is NULL.}

\item{attrition}{A vector of attrition rate at all possible
measurement occasions. The default is NULL.}

\item{...}{Other arguments, currently unused.}
}
\value{
Returns an object of class "seedpool",
	which describes the design pool.
	It contains the following information for
	each design:
\enumerate{
  \item Response patterns (0 = complete, 1 = missing);
  \item Attrition rates;
  \item Proportions of each pattern;
  \item Sample sizes.
}
}
\description{
This is the first step of SEEDMC. It generates a design pool.
This step returns the response patterns of the possible designs,
proportion of each pattern, and sample size for each design.
}
\examples{
patC <- designPool(type ="com", time = 5,
			traj = "linear", budget = 100000,
			unitcost = 20, nfix = NULL)
summary(patC)
patC$Complete$patterns
patC$Complete$patterns_notation

patM <- designPool(type ="miss", time = 6,
			traj = "linear", budget = 100000,
			unitcost = 20, attrition = rep(0.5, 6))
summary(patM)
patM$Missing$patterns
patM$Missing$patterns_notation

patQ <- designPool(type ="both",  time = 5,
			traj = "quadratic", budget = 100000, unitcost = 20)
summary(patQ)
}
\references{
Wu. W., Jia, F., Rhemtulla, M., & Little, T. D.
			(revise and resubmit). Search for efficient complete and
			planned missing data designs for analysis of change.
			Behavioral Research Methods.
}

