% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelMplus.R
\name{modelMplus}
\alias{modelMplus}
\title{Generate Mplus syntax for hypothesized model}
\usage{
modelMplus(fl = NULL, latcov = NULL, latmean = NULL, resvar = NULL, ...)
}
\arguments{
\item{fl}{The hypothesized factor loadings matrix
for a latent growth curve model.
Rows indicate indicators/repeated measures,
and columns represent latent intercept and slope(s).}

\item{latcov}{The hypothesized variance-covariance matrix of
the latent intercept and slope(s).}

\item{latmean}{The hypothesized mean vector of
the latent intercept and slope(s).}

\item{resvar}{The residual variances of the indicators.
If \code{resvar} is a scalar, then the
residual variances are the same across time.
It could also be a vector that contains the
residual variances for each of the time points.}

\item{...}{Other arguments, currently unused.}
}
\value{
Returns an object of class "seedmod",
which contains the following information.
\enumerate{
  \item Population values of model parameters.
  \item Syntax that can be used to define the
			population model in Mplus.
}
}
\description{
This is the second step of SEEDMC.
It creates the Mplus syntax for the hypothesized model,
based on the user-specified latent variance-covariance
matrix, latent means, and residual variances.
Currently, it only accommodates latent linear or quadratic
growth curve models.
}
\examples{
### Linear
fl <- matrix(NA, 5, 2)
fl[,1] <- 1
fl[,2] <- 0:4
latcov <- matrix(NA, 2, 2)
diag(latcov) <- c(28.776, 8.201)
latcov[1, 2] <- 1.56
latcov[2, 1] <- 1.56
latmean <- c(39.457, 8.063)
resvar <- 30

modL <- modelMplus (fl = fl, latcov = latcov,
			latmean = latmean, resvar = resvar)
modL

#### Quadratic
fl <- matrix(NA, 5, 3)
fl[,1] <- 1
fl[,2] <- 0:4
fl[,3] <- fl[,2]^2
latcov <- matrix(NA, 3, 3)
diag(latcov) <- c(107.08, 24.60, 1.22)
latcov[1, 2] <- -3.69
latcov[1, 3] <- -1.36
latcov[2, 3] <- -4.96
latmean <- c(13.97, -1.15, 0.2)
resvar <- c(rep(41.98,4),40)

modQ <- modelMplus(fl = fl, latcov = latcov,
		latmean = latmean, resvar = resvar)
modQ
}
\references{
Wu. W., Jia, F., Rhemtulla, M., & Little, T. D.
			(revise and resubmit). Search for efficient complete and
			planned missing data designs for analysis of change.
			Behavioral Research Methods.
}

