% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seedmcMplus.R
\name{seedmcMplus}
\alias{seedmcMplus}
\title{Search for efficient designs based on design pool and hypothesized model}
\usage{
seedmcMplus(pattern = NULL, model = NULL, nreps = 5000, seed = NULL,
  ...)
}
\arguments{
\item{pattern}{The "seedpool" object created by
the \code{designPool()} function.}

\item{model}{The "seedmod" object created by
the \code{modleMplus()} function.}

\item{nreps}{The number of replications used in the
Monte carlo approach for each design. The default is 5000.}

\item{seed}{A number used as the seed in Mplus.
If not specified, a random number will be assigned.}

\item{...}{Other arguments, currently unused.}
}
\value{
Returns an object of class "seedfoud",
which contains the following information.
\enumerate{
  \item Information in the given "seedpool" object.
  \item Information in the given "seedmod" object.
  \item Information from the Mplus output
		for all designs in the design pool,
		including the model results and messages
		about completion error for each replication.
}
}
\description{
This is the third step of SEEDMC.
Based on the "seedpool" and "seedmod" objects
generated from the first and second steps.
This function automates the Mplus Monte Carlo approach
to evaluate the efficiency of each design.
Besides the resulting R object,
a directory named "seedmc_mplus_output_yyyy-mm-dd_hh-mm-ss"
will be created under the working directory.
Here "yyyy-mm-dd_hh-mm-ss" reflects system time.
All Mplus output files are saved in the directory.
}
\examples{
patQ <- designPool(type ="both",  time = 5,
			traj = "quadratic", budget = 100000, unitcost = 20)

fl <- matrix(NA, 5, 3)
fl[,1] <- 1
fl[,2] <- 0:4
fl[,3] <- fl[,2]^2
latcov <- matrix(NA, 3, 3)
diag(latcov) <- c(107.08, 24.60, 1.22)
latcov[1, 2] <- -3.69
latcov[1, 3] <- -1.36
latcov[2, 3] <- -4.96
latmean <- c(13.97, -1.15, 0.2)
resvar <- c(rep(41.98,4),40)
modQ <- modelMplus(fl = fl, latcov = latcov,
		latmean = latmean, resvar = resvar)

seedmcoutQ <- seedmcMplus(pattern = patQ, model = modQ, nreps = 2)
#seedmcoutQ$results$Complete
seedmcoutQ$results$Complete[[1]]
# For presentation purpose, we set nreps = 2.
# In a real study, nrep should be a large number (e.g. 5000).
}
\references{
Wu. W., Jia, F., Rhemtulla, M., & Little, T. D.
			(revise and resubmit). Search for efficient complete and
			planned missing data designs for analysis of change.
			Behavioral Research Methods.
}

