% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topDesigns.R
\name{topDesigns}
\alias{topDesigns}
\title{Find the most efficient designs}
\usage{
topDesigns(x = NULL, param = NULL, out = NULL, maxerr = 0.05,
  recut = 0, acrosscut = NULL, ...)
}
\arguments{
\item{x}{The "seedfound" object obtained from the
\code{seedmcMplus()} function.}

\item{param}{A vector of the parameters of interests.
Here "I" indicates the intercept,
"S" represents the slope in a linear model,
and "S1" and "S2 represent the linear and qudratic sloples
in a qudratic model, respectively.
Valid prameter names include:
\itemize{
  \item "Means.I"
  \item "Means.S" for linear models only,
              or "Means.S1" and "Means.S2"
                for quadratic models.
  \item "Variances.I"
  \item "Variances.S" for linear models only,
             or "Variances.S1" and "Variances.S2"
             for quadratic models.
  \item "I.WITH.S" for linear models only,
        or "I.WITH.S1", "I.WITH.S2" and
        "S2.WITH.S1"for quadratic models.
}}

\item{out}{A character string indicating how the results
should be organized. If "com", the function will compute
relative efficiency only based on the complete data designs.
If "miss", it will only compute relative efficiency for
missing data designs. If "mix", it will compute relative efficiency
based on the entire design pool.}

\item{maxerr}{A number between 0 and 1.
Only the designs with the proportion of error/warning messages
smaller than \code{maxerr} will
be included when computing the relative efficiency.
The default is 0.05, meaning that the designs,
for which more than 5\% replications have error/warning messages,
will be omitted from the computation of relative efficiency.}

\item{recut}{A number between 0 and 1.
Only the designs whose relative efficiencies
are greater than \code{recut} will be included
in the final result. The default is 0,
meaning that all designs in the pool will be included.}

\item{acrosscut}{A number between 0 and 1, which
must be equal to or greater than \code{recut}.
When a number is specified, an additional table
will be created showing the designs whose relative efficiencies
are greater than \code{recut} across all parameters of interest.}

\item{...}{Other arguments, currently unused.}
}
\value{
For each parameter of interests,
return the relative efficiency (RE) of
each design, and the extra information.
}
\description{
This is the fourth step of SEEDMC.
It summarizes the "seedfoud" object obtained from
the \code{seedmcMplus()} function, and provides
relative efficiency measures of each design for
the latent means, variances and covariances.
}
\examples{
patQ <- designPool(type ="both",  time = 5,
			traj = "quadratic", budget = 100000, unitcost = 20)

fl <- matrix(NA, 5, 3)
fl[,1] <- 1
fl[,2] <- 0:4
fl[,3] <- fl[,2]^2
latcov <- matrix(NA, 3, 3)
diag(latcov) <- c(107.08, 24.60, 1.22)
latcov[1, 2] <- -3.69
latcov[1, 3] <- -1.36
latcov[2, 3] <- -4.96
latmean <- c(13.97, -1.15, 0.2)
resvar <- c(rep(41.98,4),40)
modQ <- modelMplus (fl = fl, latcov = latcov,
		latmean = latmean, resvar = resvar)

seedmcoutQ <- seedmcMplus(pattern = patQ, model = modQ, nreps = 2)
#seedmcoutQ$results$Complete
seedmcoutQ$results$Complete[[1]]
# For presentation purpose, we set nreps = 2.
# In a real study, nrep should be a large number (e.g. 5000).

finalQ <- topDesigns(x = seedmcoutQ,
			param = c("Means.I", "Means.S1", "Means.S2"),
			out = "mix", recut = 0, maxerr = 0.4)
}
\references{
Wu. W., Jia, F., Rhemtulla, M., & Little, T. D.
			(revise and resubmit). Search for efficient complete and
			planned missing data designs for analysis of change.
			Behavioral Research Methods.
}

